+++ test/index.test.js
@@ -1,7 +1,22 @@
 const test = require('ava')
-const observability = require('../index')
+
 const Observability = require('../lib/observability')
 
-test('Should return a singleton instance of Observability', t => {
+test('given the development env it should correctly setup the singleton', t => {
+  process.env.NODE_ENV = 'development'
+  delete require.cache[require.resolve('../index')]
+  // eslint-disable-next-line global-require
+  const observability = require('../index')
   t.true(observability instanceof Observability)
+  t.false(observability.initialized)
+})
+
+test('given the test env it should correctly setup the singleton', t => {
+  process.env.NODE_ENV = 'test'
+  delete require.cache[require.resolve('../index')]
+  // eslint-disable-next-line global-require
+  const observability = require('../index')
+  t.true(observability instanceof Observability)
+  // We auto init libs in testing
+  t.true(observability.initialized)
 })