+++ test/lib/helpers.test.js
@@ -1,6 +1,10 @@
 const test = require('ava')
 const helpers = require('../../lib/helpers')
 
+test('given no NODE_ENV it defaults to test in unit tests', t => {
+  t.is(helpers.getEnvironment({}), 'test')
+})
+
 test('given the right env getEnvironment should return the correct results', t => {
   process.env.NODE_ENV = 'node_env'
   t.is(helpers.getEnvironment({}), 'node_env')